+++ NEWS.md
@@ -4,6 +4,8 @@ Smaller changes:
 
 * `as.data.frame.summary.tableby()` has been implemented, and `print.summary.tableby()` updated accordingly. (#60)
 
+* `as.data.frame.summary.modelsum()` has been implemented, and `print.summary.modelsum()` updated accordingly. (#74)
+
 * `includeNA()` now has dots, and the factor method gained a `first=` argument. (#62)
 
 * Fixed a bug with assigning labels when some value names are unmatched. (#64)